requirements.txt
psycopg2
django==3.1.4
djangorestframework==3.12.2
drf-yasg==1.20.0
requests
django-admin-sortable2==0.7.7
django-admin-rangefilter
django-admin-numeric-filter
-e git+https://github.com/lukasvinclav/django-admin-numeric-filter.git@9bd51fd9c9309b54c3a13f9a594775a014b2695e#egg=django-admin-numeric-filter
django-fsm
django-fsm-admin
django-fsm-log